requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2020.6.20
chardet==3.0.4
idna==2.10
python-telegram-bot==13.7
pytz==2021.1
requests==2.24.0
six==1.16.0
tornado==6.1
tzlocal==2.1
urllib3==1.25.10
